Add ARM template to deploy App Service on Windows
Added a new Azure Resource Manager (ARM) template in `profile.arm.json` to deploy an App Service on Windows. The template includes schema and content version, metadata, parameters for resource group and resource names, variables for App Service Plan, and resources to create the App Service and App Service Plan. This facilitates the deployment of a .NET Core application on Azure.